requirements.txt
docx2txt
python-dotenv
pdfminer.six
transformers